requirements.txt
# Streamlit for the web app
streamlit>=1.11.0

# CrewAI for agent-based task management
crewai

# OpenAI API client for GPT-4 access
openai>=1.0.0

# dotenv for managing environment variables
python-dotenv>=0.21.0

# Pandas for data manipulation
pandas

# Requests for API calls to Alpha Vantage and News API
requests

# Statsmodels for ARIMA forecasting
statsmodels

# PyYAML for reading YAML config files
pyyaml

# Any additional dependencies CrewAI may have
pydantic

# For compatibility in tracing/logging (optional, based on code structure)
opentelemetry-api
opentelemetry-sdk

pysqlite3-binary
pymongo


